requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pytest>=7.1.3,<9.0.0
pytest-asyncio>=0.21.0
sqlalchemy>=1.4,<2.0
alembic>=1.12.0
greenlet>=1.1.0
python-dotenv>=1.0.1
paddleocr==3.1.0
paddlepaddle==3.1.0
transformers>=4.41.0
asyncpg>=0.28.0
httpx>=0.27.0
aiosqlite>=0.17.0
python-multipart>=0.0.6
pymupdf>=1.23.0
pillow>=10.0.0
numpy>=1.24.0 